Revert "drm/komeda: Consolidate struct komeda_drv allocations"
This reverts commit 6ca9cc39585d1ee18231f9be9d25e010777e1125.

The drm_device absolutely needs to be freed and recreated on every
bind-unbind cycle, so this is actually going in the wrong direction
memory-management wise, at least at this moment. We should be able to
move the mm once we drop bind/unbind and use only probe/remove (which
coincides with going all bridge-and-device-links.

Change-Id: I15ac8c4c0228faeb7dd992318c3f374a22409068